Prepare folder structure for comparison matrix
