+++ frontend/src/components/visual/ComparisonMatrix/DesktopComparisonMatrix/ScrollingSideArrow.js
