+++ frontend/src/components/visual/ComparisonMatrix/DesktopComparisonMatrix/scrollingHooks.js
