+++ frontend/src/components/visual/ComparisonMatrix/MobileComparisonMatrix/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+import React from 'react'
+
+export default () => <div>TODO: Mobile version not implemented yet</div>
